refactor(openings): rename verification.rs -> homomorphic.rs
The file holds both prover- and verifier-side bodies for additively
homomorphic batched openings (homomorphic_prove_batch /
homomorphic_verify_batch + RLC primitives). The previous "verification"
name was a leftover from an interim design where an OpeningVerification
trait was being folded in; the final design has only free helpers, so
the file should be named after what it actually contains, the
AdditivelyHomomorphic / AdditivelyHomomorphicVerifier trait pair it
targets.

Updates the mod / pub use lines in jolt-openings/src/lib.rs, the
crate-internal use in mock.rs, and the two stale path citations in
the spec. No behavior change.

Gates: workspace clippy --all-targets -- -D warnings green;
jolt-openings + jolt-dory + jolt-hyperkzg + jolt-verifier nextest
86/86; jolt-equivalence muldiv / modular_self_verify /
transcript_divergence / zkvm_proof_accepted_by_core_verifier all
pass.

Made-with: Cursor
