+++ crates/jolt-openings/src/homomorphic.rs
